Updated releng/autorelease Project: openflowplugin  e4faef067c5ed399ccdd686d26c09a0b7a6cf4b7
Cleanup RequestContextStack

Instead of exposing the limits as a tweakable API, we should specify the
limit at instantiation as an implementation-specific thing.

We document that RequestContextStack.createRequestContext() can return
null and let callers deal with that.

This change allows us to implement RpcContext throttling using a simple
counting Semaphore, which is more scalable and efficient than an
explicit queue. To do that, we turn RequestContextImpl into AbstractRequestContext,
which is subclassed as needed by the various RequestContextStack
implementations.

Change-Id: Ibbeafad6cd7f740284264cca22412e11696cede8
Signed-off-by: Robert Varga <rovarga@cisco.com>