Updated releng/autorelease Project: openflowjava  ccb7e5bda0598185f98d52ddd16e49ae4d48e5aa
BUG-3219: Fix OutboundQueue request error reporting

Original patch failed to describe how errors are reported when they
arrive in an ErrorMessage or are generated by the local system.

The only error path interface is the onFailure() method in the callback
the user specifies. This fits naturally, as we now define
DeviceRequestFailedException to carry the encapsulated message and pass
it to the callback.

Change-Id: Ie8eb62558991ceb7f6d5eb8d7cd547aecaf63f19
Signed-off-by: Robert Varga <rovarga@cisco.com>