Updated releng/autorelease Project: lispflowmapping  eac0ba6c41c995ec3253a780e1eec4a723b762d5
Bug 3250: Deactivate odl-lispflowmapping-northbound

The AD-SAL/Tomcat based northbound interface is deprecated in the
Lithium release, and will be removed in Beryllium.  Remove
odl-lispflowmapping-northbound from the user facing
odl-lispflowmapping-all feature, so that it doesn't get loaded by
default.  With this change, the total number of installed featured when
installing odl-lispflowmapping-all in an empty karaf container drops
from 92 to 70.  Also, the list of AD-SAL features loaded drops from

    odl-adsal-all
    odl-adsal-core
    odl-adsal-networkconfiguration
    odl-adsal-connection
    odl-adsal-clustering
    odl-adsal-configuration
    odl-adsal-thirdparty
    odl-adsal-northbound

to

    odl-adsal-core
    odl-adsal-thirdparty

Change-Id: I2cae37a5e7989782c03a0c4127b2e47a6e703083
Signed-off-by: Lorand Jakab <lojakab@cisco.com>