Updated releng/autorelease Project: packetcable  01ea06a6bb49d840cc3f2b3362bbdde7a5764fdc
Fixing the setting of the DSCP_TOSOverwrite value with the DSCPTOS object to write the value to correct element of the data byte array.
The setter that uses the byte value was correct and set index 1 where the other setter was setting the byte value to index 2.
Gate COPS data will need additional unit testing and really should be refactored for normalization and semantics.

Change-Id: I719f186d661b3749ed7d9f9f3bef20a7323d4910
Signed-off-by: Steven Pisarski <s.pisarski@cablelabs.com>