Updated releng/autorelease Project: packetcable  894b20e9ceec1c638a5fd09964284daad0341429
Added logic for the PacketcableProvider class to manage failed gate requests as well as successful.
This will ensure that the ODL restconf state and the plugin's will remain in sync. Previously, should a failed gate set be made then deleted, ODL/restconf would no longer display the gate info but it was impossible to create a new one with the same ID as the plugin would say it already existed.

Change-Id: I1b0e1b5540225d7b2afda47aaa7c8bf3f46c9508
Signed-off-by: Steven Pisarski <s.pisarski@cablelabs.com>