Updated releng/autorelease Project: openflowjava  a302a0cc64ab45e0dd3d504a6b0248c3f1f9e3df
BUG-3219: implement a shared stash for queue reuse

This introduces a global per-size stash for reuse of OutbountQueueImpl
objects. This allows for efficient reuse across multiple sessions, which
means we minimize stale memory use when some channels are idle.

The stash is implemented in terms of a ConcurrentLinkedDeque, which is
flushed from the tail and has soft references. We rely on
FinalizableReference to notify us when a particular queue is finished.

Change-Id: I78e654a0c3470bbd1c9274237b2be04f50459809
Signed-off-by: Robert Varga <rovarga@cisco.com>
(cherry picked from commit b95f82e5129f22e7b22352d028b0d32d744e924d)