Updated releng/autorelease Project: openflowjava  e2302c5d66552f0f1f833397b879a91f6369b80b
BUG-3229: Add support for PacketIn filtering

When we are facing an overload situation, we may need to suppress
PacketIns while still processing other messages. This patch adds the
interface for ConnectionAdapter users which they can use to enable
filtering of such packets.

The implementation ensures that we filter these packets as efficiently
as possible.

Change-Id: I431631922437af7a08eb8ece647904d487634bd9
Signed-off-by: Robert Varga <rovarga@cisco.com>