Updated releng/autorelease Project: integration  715ec029ccb7dea5df10bab755c3dec0b4c8f845
Beta version of a longevity test

openflow specific

reuses smaller values and single iterations of the hosts and switch
scale tests

this is intended to help meet Lithium deliverables for the integration
group, however this is also a Linux Foundation summer intern project.
I hope this can be a building block for that project.

Change-Id: If84ed95106bbde84f50edbd4c45b189341935407
Signed-off-by: Jamo Luhrsen <james.luhrsen@hp.com>