Updated releng/autorelease Project: integration  0bac77d92c319761908291d3f94863040c9ff2ff
Flip x and y axes in monitor.py on-screen table.

Flip the x and y axes in the on-screen table
shown by cluster monitor.py, so that nodes are
on the vertical axis while shards are across.

This is to make the output more readable when
there is a large number of nodes (e.g. > 10)
in the cluster vs. the typically small number
of shards.

Change-Id: If866d1409bf143391d7f020fa72a82d03be456df
Signed-off-by: Gary Wu <Gary.Wu1@huawei.com>