Updated releng/autorelease Project: controller  34e38a415bc299403657315a5b61afd432dcbbee
Use TransactionCommitOperationTimeout when commiting a Local transaction

This is to make this similar to how we do a remote transaction. The transaction
commit timeout is much larger than the operation timeout and is more appropriate
for committing because the latencies involved in committing are much larger as
committing involves replication which could sometimes be pretty slow.

Change-Id: Ie76e05bf875b4ea5eee9618519d7167f50671c8f
Signed-off-by: Moiz Raja <moraja@cisco.com>
(cherry picked from commit 3ca4e8353a32a68ebf71e98fed1954af5c419ca3)