Updated releng/autorelease Project: ovsdb  94247d07e10de9c4340de771e33c11bd9de70295
Bug 3310 Bug 3316 Use iid stamp if we configured a connection, otherwise systemid

This has been tested, and passes  SB IT tests (all 21)
As well as manually tested beyond their coverage.

PatchSet3: Switch to using ovsdb://uuid/<uuid> for default
NodeKey for OPER nodes we discover that we did not configure.

Change-Id: I270a64e13a6fca257b49d2e2ff34f754698c2df4
Signed-off-by: Ed Warnicke <hagbard@gmail.com>