Updated releng/autorelease Project: bgpcep  d9a3fe3f83d3ca7bccdb543e02249a0c88a3fa95
BUG-3348 : fixed Export Policy

Export policy located in LocRib was not informed
about peer change, therefore it didn't create any
groups. As a result, none of the routes could be
propagated to other peers. Change the LocRib to
listen to the same path as Effective RIB to obtain
also peer role changes.

Change-Id: Ic55cbba03952fbdcf7f63b76d159502c9af733b0
Signed-off-by: Dana Kutenicsova <dkutenic@cisco.com>