Updated releng/autorelease Project: integration  72f5e2394481bd35e5dd6dc03ff65a16163167bd
Using WUKS for initial node disco to avoid timing issue

also trimmed other operational checks from 60s to 6s.  for
openflowplugin, the default stats polling time is 3s, so if
it's not there in double that (6s) I think we have a real
problem and can just fail the test case then.

Change-Id: Iad30a6e582e28b5c57837159218013296c394014
Signed-off-by: Jamo Luhrsen <james.luhrsen@hp.com>