Updated releng/autorelease Project: bgpcep  1d6e02186ac85eff5734db59d3cbf5017214ddce
Fix EffectiveRibInWriter overwriting the entire route table

The fall-through in SUBTREE_MODIFIED case means that for any new write we
replace the entire route table, which has the effect of LocRib having to
run route selection even on routes which have not changed.

Fix this by removing the fall-through and deal with individual routes as appropriate.

Change-Id: I847936025d21e63c91945380744832f7af16b3dc
Signed-off-by: Dana Kutenicsova <dkutenic@cisco.com>