Updated releng/autorelease Project: openflowplugin  894c3b0f5c179fd1bb54ce6eedd74e3e5def4123
Bug:3048-Handle duplicate nodeRemoved and nodeConnectorUpdated notifications to prevent tx failures

NodeRemoved notifications are received twice by the InventoryManager, in quick successions
If the duplicate node removal operation is clubbed with other node-removal operations in transaction chain, it fails the whole tx chain even
though we retry one time.

This patch caches the deleted nodes in-order to ignore the duplicate notifications. The cache entries are expired after 10s, to reduce the memory consumption.

Signed-off-by: Kamal Rameshan <kramesha@cisco.com>