Updated releng/autorelease Project: openflowplugin  167dc20143898e3fc806ee6d541972f1e9b477ee
Use a simple lookup table for prefix conversion

This is a clear CPU/memory trade-off, where we have a pre-allocated
bytearray for valid prefixes. We then lookup the appropriate offset
into this array when performing a mask operation. If the user requests
the actual array, we return a Arrays.copyOf(), cutting just the
appropriate bytes out. Memory cost is around 2K.

Change-Id: Idd679084bd3be4a31331a38ecf978b07d738f495
Signed-off-by: Robert Varga <rovarga@cisco.com>