Updated releng/autorelease Project: ovsdb  fd1b45b2100fb2bfe7a7b58295cebdc949833680
Bug 3379:java.lang.IllegalArgumentException: ControllerEntryKey

Problem: OVSDB Controller updates are indpendent of Bridge updates but the ovsdb.yang model includes the controller data in the OVsdbBridgeAugmentation. This existing code assumed that the Controller and Bridge updates would be together and thus would always have the OvsdbBridgeAugmentation to update. That is not the case since the Controller updates can be received without a Bridge update.

Solution: Add the ControllerEntry to the OvsdbBridgeAugmentation using the Bridge if it was included with the Controller update. Otherwise, find all the OvsdbBridgeAugmentations for the connection and update them if they are using the updated Controller.

Patch also includes updates to SouthboundIT and NetvirtIT. NetvirtIT was needed so that the isConnected=true case would be tested. Southbound does not include the openflowplugin but Netvirt does. The openflowplugin is needed to test when the switch connects to the openflow controller.

Other fixes were made to the existing testOvsdbBridgeControllerInfo(). That test was using the ovsdb node ip:port as the controller target but it should have been the controller's ip:port.

Change-Id: I30785dbdf955614ad6fc675507da048434c2485e
Signed-off-by: Sam Hague <shague@redhat.com>