Updated releng/autorelease Project: integration  ea8419aaafb07e6599cfcf26b4b0f5ddaf88c423
Changing the way the mac addrs is extracted from OVS switches.

Since the "s1" interface and "s1" is hard-coded, this is a safer approach.

On Fedora, I found that the mac addr wasnt being extracted(but worked for Ubuntu).

This command will work for both platforms, and others as it uses
regular expression.

Change-Id: I008b1a3681eabdace6fbe26346750fb3b1169f7d
Signed-off-by: Kailash Khalasi <kkhalasi@iix.net>