Updated releng/autorelease Project: integration  a74619d04386c4f7b36245a6b09d8819678f8fdb
Adding a testplan file for running datastore tests in CI.

This testplan file  with be for the clustering datastore tests.
Also adding a cluster-deployer template with a large number of features
for use in exploratory testing. This template will not be used in CI.

Change-Id: Ib7117215492b34b2f37f70d362ac950754e8ef0e
Signed-off-by: Phillip Shea <phillip.shea@hp.com>
(cherry picked from commit 641cb1c25245ad8734af5e51d3a620eab40cfd98)