Updated releng/autorelease Project: integration  79c3ca4653af8729eeee9c1d6870e64ed8eb07a5
skip clear-actions check element in flow compare

the newer flow test suites has some flows that does a clear-actions
but that does not show up in the switch, as it just clears out all
existing actions.  So that element will not show up in operational
so we can skip it.

Change-Id: Iced26004df2c40c88ead5c92fa777d7178723d16
Signed-off-by: Jamo Luhrsen <james.luhrsen@hp.com>