Updated releng/autorelease Project: packetcable  4061a4c459e61b3eb22b77a5e43495d90a9dbada
Moved CMTS emulator to its own module so it can either be run standalone or be used for unit testing for other modules.

Change-Id: Id090efb3bd3ef674fc927d1820ec0f0330a002b8
Signed-off-by: Steven Pisarski <s.pisarski@cablelabs.com>

Moving fix by Tony Tkacik in the stable/lithium branch which caused an error with the integration tests during installation of the packetcable-provider bundles where it would sometimes start before MD-SAL.
Added packatcable models to required capabilities.

Change-Id: I1c6b69caae50b4f665929a82e1bb4acbd807ecec
Signed-off-by: Steven Pisarski <s.pisarski@cablelabs.com>