Updated releng/autorelease Project: openflowplugin  05d964944576807810ef9e8622f56eb2cff4c4ce
Bug 3817: Fix action convertor for OF 1.3.

The action convertor for flow entries of OpenFlow 1.3 wrongly used
instance builders for UDP source port in the conversion for the UDP
destination port. In the result, the OpenFlow plugin failed to install a
flow entry which has an action modifying UDP destination port.

Also, the action convertor wrongly used Icmpv4Type.class to create an
action modifying ICMPv4 code filed. And, the size of mask for ICMPv4
code and type fields was wrongly set into 0x0f, though the size of these
fields were 8 bits.

This patch fixes these issues so that the OF plugin can correctly
install flow entries which modify UDP destination port and ICMPv4
fields.

Change-Id: I8e972233bbea5741bb8bca39cbd281b6c00b763b
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>