Updated releng/autorelease Project: aaa  8f97e87daccb3f9f9eb8bec620ff2088fef10826
Bug 3820 Incorrect database initialization

Added username and password parameters to database initialization.  This is
necessary in order to access the database from a psql client.  This change
enforces the Singleton design pattern for IdmLightApplication.  The
getDbConnect() functionality included in UserStore, RoleStore, GrantStore
and DomainStore is consolidated in
IdmLightApplication.getConnection(Connection existingConnection) function.

Change-Id: Ib001e05548acf401c4633712cd7ab3ff6a2d2b44
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>