Updated releng/autorelease Project: integration  679fabd0e50ff8c40105c77e073e3adc73842218
creating a new test suite to find out maximum number of switches connected

the idea is not to start and stop mininet with different paramentes, but to
add (and remove) switches to the running topology

the other idea is to make that test not as long as the one running from
../../../testplans/openflowplugin-scalability-daily.txt (1.5h). The new suite
has to last much less.

Existing testplans were updated to point on certain files, not the whole directory,
so new robot file can be placed there.

Change-Id: Ifc7c9beb103b87d049f112ac57215705b8364412
Signed-off-by: Peter Gubka <pgubka@cisco.com>
(cherry picked from commit ec17f348f7557f22a3af9101140c236b6de83ecc)